backend/requirements.txt
# Core Framework
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0

# Security
django-axes==8.3.1
django-two-factor-auth==1.18.1
django-otp==1.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0

# Payments (PayFast)
dj-payfast==0.1.11
requests==2.32.5

# API & Filtering
django-filter==25.2
drf-spectacular==0.29.0

# Database
psycopg2-binary==2.9.11

# Transaction State Management
django-model-utils==5.0.0

# Audit & History
django-simple-history==3.11.0

# AWS S3 Storage
boto3==1.42.65
botocore==1.42.65
django-storages==1.14.6
Pillow==12.1.1
s3transfer==0.16.0

# PDF Invoices / Receipts
WeasyPrint==68.1

# Production Server
gunicorn==25.1.0
whitenoise==6.12.0

# Testing
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==7.0.0
factory_boy==3.3.3
coverage==7.13.4
Faker==40.8.0

# Utilities & Supporting Packages
APScheduler==3.11.2
arrow==1.4.0
asgiref==3.11.1
attrs==25.4.0
blessed==1.33.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
cssselect2==0.9.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-formtools==2.5.1
django-guardian==3.3.0
django-phonenumber-field==8.4.0
django-picklefield==3.4.0
django-q2==1.9.0
django-ratelimit==4.1.0
django-safedelete==1.4.1
fonttools==4.62.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
qrcode==8.2
redis==3.5.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
zopfli==0.4.1